Refactor to concern
